Updated InfoWindow
Included photo to the custom info window of the place shown and reorganised the structure.

Updated the output of types to allow shown ones that are contained from the 'findUILocationSwitchesTurnedOn' func from PlacesManager class to show relevant types.

Note:
Need to work on the info window for when the place name is too long, it pushes the text down making an overlap of "types" text and the UI button.